Support PING messages without final parameter ":"-prefix
The [spec] for message parameters specifies that the final parameter
may be prepended by a ":" in order signify that it should be
interpreted as one parameter, even if it contains spaces (which
otherwise delimits parameters).

However, this colon-prefix is optional if the message is unambiguous
and therefore can be left out.

As worded by the [spec]:
> Parameters that contain spaces, are empty, or begin with a ':'
> character MUST be sent with a preceding ':'; in other cases the use of
> a preceding ':' on the final parameter is OPTIONAL.

This commit therefore attempts to find a colon in the PING command and
otherwise falls back to the first space.

[spec]: https://modern.ircdocs.horse/#parameters